fix(workflow): 修正分页响应结果获取方式及空值返回逻辑
- 将分页查询空用户ID时的返回值由PageResponse.empty()改为PageResponse.success(null)
- 调整PageResponse.of方法签名，统一使用long类型参数
- 修正获取分页结果数据时，从getList()改为getData()以匹配实际数据字段
- 优化部分分页参数类型和范围的处理逻辑，提升代码一致性
- 统一委托及用户日志分页接口空值返回方式，避免异常情况

+++ ydsz-pmis-backend/ydsz-pmis-common/ydsz-pmis-common-core/src/main/java/com/njydsz/pmis/common/core/response/PageResponse.java
@@ -225,8 +225,8 @@ public class PageResponse<T> extends BaseResponse<T> {
      * @return 分页响应对象
      */
     @SuppressWarnings("unchecked")
-    public static <T> PageResponse<T> of(List<T> list, long total, int pageNum, int pageSize) {
-        return success(total, (long) pageNum, (long) pageSize, (T) list);
+    public static <T> PageResponse<T> of(List<T> list, long total, long pageNum, long pageSize) {
+        return success(total, pageNum, pageSize, (T) list);
     }
 
     /**
